+++ firmware/irremoteint.h
@@ -210,11 +210,9 @@
 
 // information for the interrupt handler
 typedef struct {
-//  uint8_t recvpin;           // pin for IR data from detector
   uint8_t rcvstate;          // state machine
-//  uint8_t blinkflag;         // TRUE to enable blinking of pin 13 on IR processing
-  unsigned int timer;     // state timer, counts 50uS ticks.
-  unsigned int rawbuf[RAWBUF]; // raw data
+  uint16_t timer;     // state timer, counts 50uS ticks.
+  uint8_t rawbuf[RAWBUF]; // raw data
   uint8_t rawlen;         // counter of entries in rawbuf
 } 
 irparams_t;
